pyproject.toml
[project]
name = "x402-facilitator-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eth-account>=0.13.7",
    "fastapi>=0.120.2",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "x402>=0.2.1",
]
